add scheme folder
