+++ server/src/Connection/Scheme/Wall.cs
@@ -1,0 +1,8 @@
+using System.Text.Json.Serialization;
+
+namespace Thuai.Server.Connection;
+public record Wall
+    {
+        [JsonPropertyName("position")]
+        public Position Position { get; init; } = new();
+    }
